requirements.txt
# 运行主流程必需依赖
requests>=2.31,<3
PyYAML>=6.0,<7

# 测试与回归验证
pytest>=8.0,<9
pytest-cov>=5.0,<6
